Feat: Envfile loader (#8)
* Build .env file loader

* Test case: invalid .env file format

* Build more test cases for env file loading

* Add key validation and refine test fixture.

For now, keys are invalid if they contain any whitespace.

Switched the test fixture to a r-string as I considered that reading a
`.env` file with a key `FOO=BAZ\tBIZ` would not have the expectation
that the `\t` would be evaluated.

* Add tests around stripping quotes from values

Leading and trailing single or double quotes should be stripped from
values when read. Inner quotes should be ignored. If there is a mismatch
of the first and last character quote style the quotes should not be
removed.