uv workflow (#13)
* Move requirements into pyproject.toml

Use dependency groups to separate the various requirements. Dropping the
`types-requests` requirement as that looks to be a mistake.

* Switch to hatchling build backend

* fix: add src path for macos gha coverage

* create .python-versoin pin file

Pin version to 3.10, the oldest supported version for development

* create uv.lock file

* update noxfile for uv

Replaces all existing sessions with uv specific sessions. Also
simplifies all session use.

* Update documentation for uv requirements

* update gha to use new nox sessions

This change also updates the version of all external gha used in the ci.

* Update pre-commit-ci hooks

It is no longer required to use pre-commit locally. The hooks selected
are self-healing and allow pre-commit-ci to support any of the nits.