Mutate `os.environ` with Eggviron actions (#17)
* only run loader once

The previous implementation would run a loader twice when `self_strict`
was True.

* Add `mutuate_environ` flag, new default behavior

The default behavior of an Eggviron instance will be to mutate the
`os.environ` values as Eggviron `__setitem__()` is called or `load()` is
called. For loaders, the `os.environ` is updated after *each* loader
provided.

This will allow the dev to load a `.env` file with values needed for
additional loaders. As an example: using the `.env` file to provide the
secrets needed to use the AWS Parameter Store loader.